[IMP] add messages delete return

+++ services/account.js
@@ -114,15 +114,20 @@ exports.quickTransaction = async (id, cibleId, userId, sum) => {
     return account
 }
 
-exports.deleteAccountByID = (id, userid) => {
+exports.deleteAccountByID = async (id, userid) => {
+    const verifAccount = await this.getAccountById(id)
     //sera à améliorer lorsque les droits d'accés seront créés (nécéssitant droit de supression D)
     deleteRights=true
     if(deleteRights){
-        return account.destroy({
-            where: {
-                id
-            }
-        })
+        if(verifAccount){
+            return account.destroy({
+                where: {
+                    id
+                }
+            })
+        }else{
+            throw new NotFound("this account doesn't exist")
+        }
     }else{
         throw new NotLogged("you haven't access rights")
     }
